[kernel][refactor] Move pmem to kernel
